+++ Fontes/ACBrDFe/ACBrCTe/ACBrCTe.pas
@@ -114,13 +114,13 @@ type
 
     procedure SetStatus(const stNewStatus: TStatusACBrCTe);
 
-    function Enviar(ALote: Integer; Imprimir: Boolean = True;
+    function Enviar(ALote: Int64; Imprimir: Boolean = True;
       ASincrono: Boolean = False): Boolean;  overload;
     function Enviar(const ALote: String; Imprimir: Boolean = True;
       ASincrono: Boolean = False): Boolean;  overload;
 
     function Consultar( const AChave: String = ''; AExtrairEventos: Boolean = False): Boolean;
-    function Cancelamento(const AJustificativa: String; ALote: Integer = 0): Boolean;
+    function Cancelamento(const AJustificativa: String; ALote: Int64 = 0): Boolean;
     function EnviarEvento(idLote: Int64): Boolean;
     function Inutilizar(const ACNPJ, AJustificativa: String;
       AAno, ASerie, ANumInicial, ANumFinal: Integer): Boolean;
@@ -712,7 +712,7 @@ begin
   end;
 end;
 
-function TACBrCTe.Enviar(ALote: Integer; Imprimir: Boolean = True;
+function TACBrCTe.Enviar(ALote: Int64; Imprimir: Boolean = True;
       ASincrono: Boolean = False): Boolean;
 begin
   Result := Enviar(IntToStr(ALote), Imprimir, ASincrono);
@@ -787,7 +787,7 @@ begin
   Result := True;
 end;
 
-function TACBrCTe.Cancelamento(const AJustificativa: String; ALote: Integer): Boolean;
+function TACBrCTe.Cancelamento(const AJustificativa: String; ALote: Int64): Boolean;
 var
   i: Integer;
 begin
