+++ Fontes/ACBrDFe/ACBrNFe/ACBrNFe.pas
@@ -99,7 +99,7 @@ type
       sMensagem: TStrings = nil; sCC: TStrings = nil; Anexos: TStrings = nil;
       StreamNFe: TStream = nil; const NomeArq: String = ''; sReplyTo: TStrings = nil); override;
 
-    function Enviar(ALote: integer; Imprimir: Boolean = True;
+    function Enviar(ALote: Int64; Imprimir: Boolean = True;
       Sincrono: Boolean = False; Zipado: Boolean = False): Boolean; overload;
 
     function GetNomeModeloDFe: String; override;
@@ -113,7 +113,7 @@ type
 
     function Enviar(const ALote: String; Imprimir: Boolean = True;
       Sincrono: Boolean = False; Zipado: Boolean = False): Boolean; overload;
-    function Cancelamento(const AJustificativa: String; ALote: integer = 0): Boolean;
+    function Cancelamento(const AJustificativa: String; ALote: Int64 = 0): Boolean;
     function Consultar(const AChave: String = ''; AExtrairEventos: Boolean = False): Boolean;
     function EnviarCartaCorrecao(idLote: Int64): Boolean;
     function EnviarEvento(idLote: Int64): Boolean;
@@ -639,7 +639,7 @@ begin
   end;
 end;
 
-function TACBrNFe.Cancelamento(const AJustificativa: String; ALote: integer = 0): Boolean;
+function TACBrNFe.Cancelamento(const AJustificativa: String; ALote: Int64 = 0): Boolean;
 var
   i: integer;
 begin
@@ -701,7 +701,7 @@ begin
   Result := True;
 end;
 
-function TACBrNFe.Enviar(ALote: integer; Imprimir: Boolean = True;
+function TACBrNFe.Enviar(ALote: Int64; Imprimir: Boolean = True;
   Sincrono: Boolean = False; Zipado: Boolean = False): Boolean;
 begin
   Result := Enviar(IntToStr(ALote), Imprimir, Sincrono, Zipado);
